- Add feature for get user lotto numbers from Jsoup(HTML parser) use to lottery web site.

+++ app/src/main/java/com/jeepchief/newlotterycheck/model/rest/LotteryDTO.kt
@@ -26,7 +26,7 @@ data class LotteryDTO(
 
     @Expose
     @SerializedName("firstWinamnt")
-    val firstWinAmnt: Int,
+    val firstWinAmnt: Long,
 
     @Expose
     @SerializedName("drwtNo1")
